Moves strategies into separate namespace
Moves all the strategies into a different namespace
(`OctoMerge::Strategy`).
