+++ lib/octo_merge/strategy.rb
@@ -1,0 +1,11 @@
+require "octo_merge/strategy/base"
+
+require "octo_merge/strategy/merge_with_rebase"
+require "octo_merge/strategy/merge_with_rebase_and_message"
+require "octo_merge/strategy/merge_without_rebase"
+require "octo_merge/strategy/rebase"
+
+module OctoMerge
+  module Strategy
+  end
+end
