+++ lib/octo_merge/strategy/merge_without_rebase.rb
@@ -1,0 +1,17 @@
+module OctoMerge
+  module Strategy
+    class MergeWithoutRebase < Base
+      def run
+        git.checkout(master)
+        git.fetch(upstream)
+        git.reset_hard("#{upstream}/#{master}")
+
+        pull_requests.each do |pull_request|
+          git.remote_add("#{pull_request.remote} #{pull_request.remote_url}")
+          git.fetch(pull_request.remote)
+          git.merge_no_ff("#{pull_request.remote}/#{pull_request.branch}")
+        end
+      end
+    end
+  end
+end
