+++ spec/octo_merge/strategy/merge_with_rebase_spec.rb
@@ -1,21 +1,20 @@
 require 'spec_helper'
 
-describe OctoMerge::MergeWithoutRebase do
+describe OctoMerge::Strategy::MergeWithRebase do
   include SetupExampleRepos
 
   let(:expected_history) do
 <<-'EXPECTED_HISTORY'.strip
-*   Merge remote-tracking branch 'bob/sunglasses'
+*   Merge branch 'sunglasses'
 |\
 | * Adds sunglasses
-* |   Merge remote-tracking branch 'alice/cowboy_hat'
-|\ \
-| * | Adds cowboy_hat
-* | | Adds piercing
-| |/
-|/|
-* | Adds tattoo
 |/
+*   Merge branch 'cowboy_hat'
+|\
+| * Adds cowboy_hat
+|/
+* Adds piercing
+* Adds tattoo
 * Adds earrrings
 EXPECTED_HISTORY
   end
